+++ net/socket/socket_pool_additional_capacity.h
@@ -31,7 +31,7 @@ class NET_EXPORT_PRIVATE SocketPoolAdditionalCapacity {
   static SocketPoolAdditionalCapacity Create();
 
   static SocketPoolAdditionalCapacity CreateForTest(double base,
-                                                    int capacity,
+                                                    size_t capacity,
                                                     double minimum,
                                                     double noise);
 
@@ -39,15 +39,15 @@ class NET_EXPORT_PRIVATE SocketPoolAdditionalCapacity {
   // `sockets_in_use` should be counted pre-allocation and `socket_soft_cap`
   // is likely being passed down from `g_max_sockets_per_pool`.
   SocketPoolState NextStateBeforeAllocation(SocketPoolState current_state,
-                                            int sockets_in_use,
-                                            int socket_soft_cap) const;
+                                            size_t sockets_in_use,
+                                            size_t socket_soft_cap) const;
 
   // Calculates the next `SocketPoolState` after the release of a socket.
   // `sockets_in_use` should be counted post-release and `socket_soft_cap`
   // is likely being passed down from `g_max_sockets_per_pool`.
   SocketPoolState NextStateAfterRelease(SocketPoolState current_state,
-                                        int sockets_in_use,
-                                        int socket_soft_cap) const;
+                                        size_t sockets_in_use,
+                                        size_t socket_soft_cap) const;
 
   explicit operator std::string() const {
     return base::StringPrintf(
@@ -67,29 +67,30 @@ class NET_EXPORT_PRIVATE SocketPoolAdditionalCapacity {
   static void LogStateTransition(SocketPoolAction action,
                                  SocketPoolState current_state,
                                  SocketPoolState next_state,
-                                 int sockets_in_use);
+                                 size_t sockets_in_use);
 
   SocketPoolAdditionalCapacity() = default;
   SocketPoolAdditionalCapacity(double base,
-                               int capacity,
+                               size_t capacity,
                                double minimum,
                                double noise);
 
   // Helper for NextStateBeforeAllocation to avoid duplicate logging code.
   SocketPoolState NextStateBeforeAllocationImpl(SocketPoolState current_state,
-                                                int sockets_in_use,
-                                                int socket_soft_cap) const;
+                                                size_t sockets_in_use,
+                                                size_t socket_soft_cap) const;
 
   // Helper for NextStateAfterRelease to avoid duplicate logging code.
   SocketPoolState NextStateAfterReleaseImpl(SocketPoolState current_state,
-                                            int sockets_in_use,
-                                            int socket_soft_cap) const;
+                                            size_t sockets_in_use,
+                                            size_t socket_soft_cap) const;
 
   // This helper function for `NextStateBefore(Allocation|Release)` handles
   // common logic. Returns a SocketPoolState if the common logic is controlling,
   // and std::nullopt otherwise.
-  std::optional<SocketPoolState> NextStateCommonImpl(int sockets_in_use,
-                                                     int socket_soft_cap) const;
+  std::optional<SocketPoolState> NextStateCommonImpl(
+      size_t sockets_in_use,
+      size_t socket_soft_cap) const;
 
   // Unlike other functions in this class, this one will CHECK on invalid
   // constants and inputs. As such, all validation must be performed before we
@@ -99,12 +100,13 @@ class NET_EXPORT_PRIVATE SocketPoolAdditionalCapacity {
   // amount of `capacity_` already allocated for `NextStateBeforeAllocationImpl`
   // and the amount of `capacity_` free for `NextStateAfterReleaseImpl`. This is
   // done to ensure the probability converges toward 1 correctly for each.
-  bool ShouldTransitionState(SocketPoolAction action, int actions_taken) const;
+  bool ShouldTransitionState(SocketPoolAction action,
+                             size_t actions_taken) const;
 
   // See the implementation of `ShouldTransitionState` for how these constants
   // are used and bound in calculating the probability of a state transition.
   double base_ = 0.0;
-  int capacity_ = 0;
+  size_t capacity_ = 0;
   double minimum_ = 0.0;
   double noise_ = 0.0;
 };
